fixes #24.
By adding the following to the sdkconfig.h in the project using framework=espidf the framework-arduinoesp32 code is automatically included/compiled.

The sdkconfig.h is checked for CONFIG_AUTOSTART_ARDUINO and CONFIG_ENABLE_ARDUINO_DEPENDS which are defined by the arduino-esp32 Kconfig.

Also fixed an issue related to CONFIG_HEAP_POISONING_LIGHT leading to link errors (this is the default option from arduino-esp32)