Better fix for esp32_exception_decoder missing space
Recent versions of ESP-IDF have a bug in which the Backtrace: line is
malformed.

Old/Good: Backtrace: 0xN:0xM 0xN:0xM 0xN:0xM ...
New/Bad:  Backtrace:0xN:0xM0xN:0xM 0xN:0xM ...

I issued https://github.com/espressif/esp-idf/pull/9138 against
ESP-IDF to fix the underlying problem, but it is unclear when or if
that PR will be accepted, especially since the monitor program in
ESP-IDF is immune to the problem.

By using a decoding technique similar to the one in ESP-IDF's monitor,
either backtrace format can be accepted.

Supersedes #687 , which does not quite work, in that it misses one of
the entries.