Fix matching of "components" dir when there's a symlink in the path to .platformio
compile_source_files() uses a relative path for the object if the source file is within the "components" tree.  It checks the path string with .startswith(), though, and that can fail if the paths point to the same file but look different due to symlinks.

In my case, it saw two different source paths in target-__idf_efuse-5f3f9f6e3513e0903efa.json:
1. /space/home/user/.platformio/packages/framework-espidf/components/efuse/esp32c6/esp_efuse_fields.c
2. /space/home/user/.platformio/packages/framework-espidf/components/efuse/src/esp_efuse_fields.c

But my home directory uses a symlink, so components_dir was:
- /home/user/.platformio/packages/framework-espidf/components

This resulted in attempting to build the same object file for both:

*** Multiple ways to build the same target were specified for: /space/home/user/git/project/esphome/.esphome/build/device/.pioenvs/device/esp_efuse_fields.c.o  (from ['/space/home/user/.platformio/packages/framework-espidf/components/efuse/esp32c6/esp_efuse_fields.c'] and from ['/space/home/user/.platformio/packages/framework-espidf/components/efuse/src/esp_efuse_fields.c'])
File "/home/user/.platformio/platforms/espressif32/builder/frameworks/espidf.py", line 773, in compile_source_files

This fix simply resolves any symlinks in components_dir before using it for the comparison.